requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
gunicorn  # For production server (optional)
requests>=2.31.0  # For HTTP API calls

# Django REST Framework for API endpoints
djangorestframework>=3.14.0
django-cors-headers>=4.0.0  # For CORS handling with React
djangorestframework-simplejwt>=5.2.0  # For JWT authentication
